requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
cffi==1.14.4
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.3.0
pyttsx3==2.90
schedule==0.6.0
six==1.15.0
toml==0.10.2
yarl==1.5.1
